refactor: conditions routes
