+++ server/src/routes/conditions/conditions.routes.ts
@@ -1,0 +1,5 @@
+import { Router } from 'express'
+import controller from './conditions.controller'
+
+export default Router()
+  .get('/conditions', controller.get)
