+++ server/src/routes/index.ts
@@ -2,8 +2,10 @@ import { Router } from 'express'
 import userRoutes from './user/user.routes'
 import userDataRoutes from './user-data/user-data.routes'
 import sessionDataRoutes from './session-data/session-data.routes'
+import spotRoutes from './conditions/conditions.routes'
 
 export default Router()
   .use(userRoutes)
   .use(userDataRoutes)
   .use(sessionDataRoutes)
+  .use(spotRoutes)
